requirements.txt
streamlit
chromadb
python-dotenv
pypdf
groq
sentence-transformers
